Fix weird https redirection downgrade bug
For some reason this endpoint occasionally sends a 302 redirect to an http image instead of https. >:(
Ideally this should be reverted back to using https at some point, but I can't reproduce this in any
browser, so I have no idea what exactly is causing it. Sometimes changing the referer works for specific
images, other times it doesn't. This isn't the most secure, but it should be the most reliable. Hopefully
whoever is mitm-ing you isn't concerned with what vsco profiles you're downloading.